fix bugs order return

+++ customer-service/src/main/java/vn/viettel/customer/controller/CustomerController.java
@@ -229,7 +229,7 @@ public class CustomerController extends BaseController {
     //Lấy danh sách customer theo danh sách id
     @RoleFeign
     @GetMapping(value = { V1 + root + "/feign-cusinfo"})
-    public List<CustomerDTO> getCustomerInfo(@RequestParam(required = false) Long status, @RequestParam List<Long> customerIds) {
+    public List<CustomerDTO> getCustomerInfo(@RequestParam(required = false) Integer status, @RequestParam List<Long> customerIds) {
         return service.getCustomerInfo(status, customerIds);
     }
